server/requirements.txt
astroid==1.6.1
Django==2.0.2
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.12
pkg-resources==0.0.0
PyJWT==1.5.3
pylint==1.9.1
pylint-django==0.11.1
pylint-plugin-utils==0.2.6
pytz==2018.3
six==1.11.0
wrapt==1.10.11
